please.sh upgrade curl: allow partial upgrade to v7.62.0 to complete
Previously, the mingw-w64-curl packages were built for version 7.62.0,
but due to a temporary DNS problem, the automated build failed to push
the update to MINGW-packages (and to continue with the MSYS packages).

Let's hack the script to let it proceed by restarting the process and
skipping the mingw-w64-* packages.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>