pacman-helper: learn about additional "extra packages"
"Extra packages" are packages that are built from the same source code
as a base package. For example, the "heimdal" source package is used to
build the "heimdal", the "heimdal-devel" and the "heimdal-libs" package.

When looking for source packages, let's just skip looking when we
process the name of a known extra package.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>