installer: remove some irrelevant Perl packages
Continuing the earlier work to reduce the Perl packages with the goal of
decreasing the size of the installer, we now skip a couple of Perl
packages when bundling the installer.

At least some of these packages are required to build and test Git for
Windows (e.g. MakeMaker for building and TAP for testing with `prove`),
therefore they cannot be flat out removed in Git for Windows' SDK.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>