installer: log the post-install output
Obviously this will work only when run with /LOG=<file>, but it is
better than what we have currently.

While at it, we now also detect the presence of the post-install.log
file correctly, and fix the bug hidden by the erroneous check where we
were fooled by a non-zero exit code: post-install.bat deletes itself, so
it will *always* have a non-zero exit code. But if it deletes itself
successfully, we know it actually succeeded.

This patch was written in response to
https://github.com/git-for-windows/git/issues/1449

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>