git-extra: add back a deprecated `git update`
In the Git for Windows project, we learned the hard way, several times,
that it is not good to just remove or rename stuff without giving the
users a chance to adjust.

In this instance, we just renamed `git update` to `git
update-git-for-windows`, and unless the user reads the release notes
(and let's be honest: only a teeny tiny fraction of the users read
those), she has no chance of knowing what to do next.

So let's be nice and add back the `git update` command; it will output
the deprecation notice and hand off to the `update-git-for-windows`
subcommand.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>