get-sources: allow for working around BinTray's let-down
BinTray let us down royally: it "rate-limited" us due to overuse, which
does not mean what you think it means: it does not mean that we now have
super-slow downloads and uploads. Instead, it means we have *no*
downloads and *no* uploads.

Since this maintainer had to get the entire set of source packages for
the just-released version 2.16.1(2), he had to re-generate the source
packages now locked away by BinTray, and this script was ill-prepared to
allow for manual copying of source packages to cached-source-packages/.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>