pacman-helper: switch to Azure Blobs
Granted, it was a nice feature in BinTray that you could upload all
files relating to a certain package's version, and then publish them all
in one fell swoop via a single REST API call.

However, since we cannot use BinTray since past Friday, with no
indication of any inclination on BinTray's side to answer my calls, we
have to switch away.

Might just as well switch to Azure Blobs, at least for the time being,
as it worked really well for Git for Windows' snapshot builds (modulo
stupid mistakes on my side, of course).

We no longer get complete transactional safety, but let's see how much
we actually need that in reality: the main purpose of the Pacman
repository is to allow users to keep their Git for Windows SDK
up-to-date, but we will most likely start to recommend cloning
https://github.com/git-for-windows/git-sdk-64 (and .../git-sdk-32)
instead, as soon as we made that work robustly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>