please.sh: refactor out code to determine package version from PKGBUILD
This code is a bit finicky because it tries to avoid sourcing the
entire PKGBUILD script in favor of using some sed expressions to extract
both pkgver and pkgrel and construct the correct package version.

The package version is needed to determine what is the current version
of a package, say, when calling `please.sh install <package>`.

Previously, there were quite a couple of breakages when we added support
for building new packages because they would build alright, but fail to
install because the sed expression could not handle *that* particular
package's PKGBUILD.

By factoring out the code, we make it easy to test this before starting
a full-fledged upgrade, so that we can have some confidence that the
upgrade won't fail at the very end.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>