please.sh: allow for upgrading perl-Net-SSLeay
This Perl module stopped working when we upgraded Perl... And as a
consequence, `git send-email` stopped working.

Not a big deal for most Git for Windows users, but it broke the
automated part of Git for Windows' release engineering (as part of every
Git for Windows release, an announcement mail is sent).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>