please.sh upgrade: learn to upgrade more Perl packages
So far, we only rebuilt Net-SSLeay because it broke due to the upgrade
to Perl v5.26.

There are more Perl packages, though, that broke because their .dll
files need to be rebuilt, e.g. Term::ReadKey (required for interactive
authentication by `git svn`, reported as issue 1488).

A manual inspection turned up a few more packages containing .dll files.
Let's just rebuild all of them. Apart from YAML::Syck (for which a new
version is available), that means that the pkgrel needs to be bumped
from 1 to 2 (except for XML-Parser, where it needs to be bumped from 2
to 3).

This fixes https://github.com/git-for-windows/git/issues/1488

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>