please.sh upload: use a lease ID to avoid concurrent uploads
In the current setup, the packages are built and uploaded in VSTS jobs,
and this maintainer needs to ensure that they are serialized lest the
uploads interfere with one another.

This is not an ideal situation. Let's add code that takes a lease (or
waits for a minute to try again, up to 2h), so that only the uploading
part gets serialized, and so that it gets serialized automatically.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>