estimate-installation-size: prefer innounp if available
It is a shame, really, that innoextract segfaults for a while now. Of
course, even this developer was not able to figure out in a quick gdb
run what is going wrong, so it is understandable.

On the other hand, innounp (which cannot be packaged easily in MSYS2
because it requires a Borland Pascal compiler) works pretty well. So
let's use it if available.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>