please.sh bundle-pdbs: fix execution from Git Bash
The intended use case for bundle-pdbs is to be run as part of the
`publish` subcommand, which installs the Git packages and therefore
cannot be run from any SDKs git-bash.exe (because that executable will
be overwritten). Therefore, this step needs to assume that it runs in a
regular Git Bash from an end-user installation of Git for Windows, in
particular it has to assume that neither xz.exe nor zip.exe are in the
PATH.

Let's make this use case work by executing the SDK versions of tar.exe
(which will then find xz.exe) and zip.exe.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>