please.sh finalize relnotes: fix --delete-existing-tag
We did delete the existing tag. But then we still continued as if that
tag had existed and increment the patch level (i.e. if, say,
v2.16.2.windows.1 was tagged still from a previously failed run, we
would delete it, but then go on to tag v2.16.2.windows.2!).

The reason was that we failed to introduce a loop in a4689179e
(please.sh finalize: instead of ignoring a colliding tag, delete it,
2018-01-22) which would try to determine the latest tag again, then
check again whether we should delete it.

This commit finally makes everything right.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>