git-extra: in Git for Windows SDKs, initialize /usr/src/* by default
The git/, build-extra/, MINGW-packages/ and MSYS2-packages/ worktrees
are pretty important in the Git for Windows SDKs. Let's initialize them
by default so that they are there (but do not fetch them just yet, just
set up core.autoCRLF=false and the `origin` remote).

This is in line with the way the SDK installer worked so far: the idea
of this commit & friends is to replace the SDK installer by a simple

	git clone https://github.com/git-for-windows/git-sdk-64 &&
	git-sdk-64/git-bash.exe

For historical reasons, we do *not* initalize those worktrees in the SDK
installer if the environment variable `JENKINS_URL` is set.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>