git-extra: show the SDK welcome only in interactive sessions
It is pointless to show it in non-interactive sessions ;-)

While at it, also disable the auto-initializing as well as the desktop
icon creation when in non-interactive sessions.

This makes the test for JENKINS_URL largely moot (which was a crux
anyway, as we simply did not have a way in the Batch script used by the
SDK installer to test whether the session is interactive).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>