please.sh bundle-pdbs: offer to unpack instead of bundling
We will want to include the .pdb files (optionally) into the installers.
Let's make that easy by optionally unpacking them into a given location
instead of bundling them into a .zip file.

We need to be careful to *not* remove the unpack directory in that case,
as it is quite possible that it has been pre-populated.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>