please.sh upgrade: copy artifacts *before* trying to upload
It is distressing when an upload fails, and even more distressing when
the artifacts that wanted to be uploaded/pushed are no longer there.

The idea of the --directory=<...> option is to have those artifacts even
when something fails later in the script, especially when running in an
automated build in the cloud. So let's make sure that the artifacts are
copied as early as possible.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>