upload-to-github: fail on server errors
The v2.17.0 release had a problem where 32-bit BusyBox-based MinGit
would not upload properly (it uploaded 100% but for some reason it would
not finish uploading).

We thought that this would be caught, but `curl` (without `--fail`) does
not produce an exit code indicating errors, and with `--fail` it would
not output anything on error!

So let's do some manual trickery here to deduce from the output of `curl
-i` (which includes the response headers in the output) whether we were
successful in uploading the file(s) or not.

The problem was reported in
https://github.com/git-for-windows/git/issues/1611

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>