sdk: set and use src_dir for the commands working on source code
The `sdk` shell function intended to help working with the Git for
Windows SDK is designed to make tasks easy such as building Git, or a
Git for Windows installer.

To that end, it auto-initializes worktrees when needed.

But we repeat an awful lot near-hardcoded paths to the worktree. And the
user will not necessarily know where the worktree is, either.

So let's store the worktree's path in the src_dir variable (which
is changed whenever we touch a different project).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>