sdk-installer: switch to cloning the Git for Windows SDK
Previously, we went through an elaborate setup, designed to work around
MSYS2's many assumptions that Git for Windows breaks.

Things became really tricky when we switched to GPG-signing the Pacman
packages, and MSYS2 really wanted to make sure that any custom GPG keys
installed into Pacman's keyrings are blown away at every possible turn.

And finally, everything really broke down when BinTray "rate limited" us
(which in reality meant that we were blocked; nobody could download
*anything* from our Pacman repository).

Even switching to Azure Blobs caused problems because the current SDK
installer still tries to access BinTray.

So we started the work recently to make the Git mirror of the Git for
Windows SDK (which is used regularly by the automated builds of Git for
Windows) more widely usable.

The idea was that the Git for Windows SDK can be cloned, and that is
that. To save bandwidth and disk space, we can do it with a shallow
clone.

And to make it very convenient for the contributors, we can even change
the SDK installer to do exactly that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>