please.sh upgrade: learn about nodejs, too
Every time mingw-w64-icu is upgraded to a new major version, node.exe
stops working (exit code 53, suggesting that a dependency could not be
found, which is true, as the version number is encoded in, say,
/mingw64/bin/libicuuc61.dll).

This is very unfortunate, as Git for Windows relies on a couple of
node.js scripts (and I would even like to see more Bash scripts to be
converted to using node.js because the latter at least allows for some
object oriented programming, and for some proper error handling via
exceptions).

So let's bite the sour apple and build mingw-w64-nodejs ourselves.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>