please.sh: do not require `cygpath` to be in the PATH
One of the uses cases for the please.sh script is to synchronize 32-bit
and 64-bit SDKs' Pacman packages. To avoid problems when upgrading Bash
(or the MSYS2 runtime), we run `please.sh sync` with a *different* Bash,
e.g. MinGit's, where no `cygpath` is available.

So let's stop relying on the `cygpath` utility, and start relying on
Bash's builtin `pwd -W`, `dirname` and `basename` instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>