shears: optionally use `git rebase -i --rebase-merges`
The patches to implement the --rebase-merges mode have been contributed
to core Git and are starting to stabilize.

In preparation for this mode (which was implemented in order to subsume
almost all of the functionality of the Git garden shears), let's modify
the shears to make use of that mode instead of running the (slow) shell
version of it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>