Vim: Source defaults.vim
Since version 8 Vim comes with the defaults.vim file distributed. That
file is already sourced on startup (as seen by the `:scriptnames`
command). However it loads usually after the system vimrc file, so it
could potentially overwrite settings that are set here without notice.

Instead, go ahead and source the file directly here and to prevent that
it will be loaded again after this file has been processed, set the
`skip_defaults_vim` variable.

While at it, reset the mouse setting which is set in defaults.vim. When
the mouse option is enabled, this prevents the usual copy/pasting with a
mouse from the terminal. Instead you would have to use <Shift> Mouse to
mark and copy. Revert this to the usual terminal behaviour.

Also do remove the last-position jump that is defined in defaults.vim,
since this file defines its own last-positioning jump so it is not
necessary to have it execute twice.

Signed-off-by: Christian Brabandt <cb@256bit.org>