Vim: Better commandline completion
The wildmode setting defines, how the files (and commands) are completed
in Vim. This means editing a file using `:e file` you can press tab and
have file autocompleted.

Unfortunately, the default command line completion for
files is rather un-intuitive when compared to e.g. bash or zsh (which
complete to the longest common string and then stop to show all possible
candidates and wait for the user to decide what to complete), since it
will complete to the first match.

This is not only unintuitevly, but also cumbersome, since you have to
watch carefully for the path being completed, possibly hit <BS> several
times to correct the wrong match, type something again, have it possibly
be completed to the wrong match again, correct again, etc...

To fix this, set the 'wildmode' option to a more sane setting, that will
stop at the longest common string and show possible completions and wait
for the user to complete further.

I argued to have this default setting changed
https://groups.google.com/d/msg/vim_dev/gL5DXac4NaU/sYTDeWctAwAJ, but
Bram decided that he likes the current behaviour better.

Signed-off-by: Christian Brabandt <cb@256bit.org>