Vim: Simplify last-position jump a bit
The last position jump that is defined here, uses a couple of specific
file names to decided whether to skip jumping to the last position.

Vim however already knows some of those filenames and will set the
filetype option for those files to 'gitcommit' or 'gitrebase'.

So simplify this logic a bit, by checking against the filetype setting
when deciding if the cursor should be positioned at the last known
position for that file. This uses Vims regular expression engine to
check against a specific filetype, e.g. the filetype option matches
either 'commit\|gitrebase'

This is also more in sync with the example in the help (and in
defaults.vim) of Vim (:h last-position-jump), which basically skips this
for all commit like filetypes (so not only git, but also e.g. for the
mercurial filetype hgcommit).

The other filenames are currently not detected automatically by Vim, so
leave them there as is. However, if those are specific files that should
be detected automatically by Vim (and set to a specific filetype) it
might make sense to send a patch upstream.

While at it, have all comments start with a '" ' (e.g. add a leading
space after the comment character) to be consistent. This is only a
stylistic change.

Signed-off-by: Christian Brabandt <cb@256bit.org>