Vim: diff files are automatically detected by Vim
Files with the '.diff' extension are automatically detected by Vim as
filetype=diff. So that specific autocommand can be removed.

Also files with the .patch extension are already detected as
filetype=patch, unless the first lines starts with 'From ...' in which
case Vim will set the filetype to gitsendemail.

Actually the logic there is:

  au BufNewFile,BufRead *.patch
	\ if getline(1) =~ '^From [0-9a-f]\{40\} Mon Sep 17 00:00:00 2001$' |
	\   setf gitsendemail |
	\ else |
	\   setf diff |
	\ endif

So leave this untouched for now and have patch files always set to
filetype diff.

Signed-off-by: Christian Brabandt <cb@256bit.org>