Vim: Document and extend the statusline setting
In addition, instead of the absolute path, show only the relative path
(%f/%F). This makes it simpler to edit another file using :e because one
knows directly from which path to complete. Also this saves a bit of
space.

While at it, drop the `%<` (truncation atom), since by default, Vim will
start truncation from the start of the statusline (:h 'stl' search for <)

If desired, the path could be shown as well.

Signed-off-by: Christian Brabandt <cb@256bit.org>