installer/portable/mingit: use xpdf-tools instead of xpdf
Previously, the xpdf package did not actually contain xpdf. Instead, it
shipped the quite useful add-on tools such as pdftotext.

However, MSYS2 decided that the xpdf package really should contain xpdf.
Which requires Qt. Which takes 6GB on disk (and 1GB to download).

Worst of all: MSYS2 decided that those useful tools should be
*excluded*, under the premise that poppler has them already (but
poppler's pdftotext is not as lightweight as xpdf's).

So we started to build the xpdf-tools package, based on MSYS2's
definition of its xpdf package. Except that we reverse their decision:
we include those tools, and we exclude xpdf, and we do not drag in Qt.

Pointed out by Beat Bolli.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>