please.sh upgrade curl: allow rebuilding only mingw-w64-curl
Usually, we upgrade both curl and mingw-w64-curl at the same time.
However, there are occasions when we fix bugs that affect only the
mingw-w64-curl package, but not the curl one. In this case, we want to
avoid packaging and uploading a totally unnecessary curl package (with
increased pkgrel).

Now we can avoid that: by passing the --only-mingw flag.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>