installer/portable/etc: *do* include PCRE2 .dll files
We tried to slim down the payload previously, but unfortunately in our
attempts, we prevented pcre2*.dll from being bundled.

This commit relaxes the exclusion list, even at the price of including
libpcre2-32-0.dll and libpcre2-16-0.dll, which are technically not
needed. But the pattern is more robust this way.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>