please.sh prerelease: offer to use another branch than `master`
For the MinGit backport releases, we will want to use a sort of "time
travel": we will use specific branches in the Git for Windows SDKs, in
build-extra and in MINGW-packages.

For simplicity's sake, we will use the same branch name in all four
repositories, e.g. mingit-2.14.x-releases for the v2.14.* release train.

This commit allows us to call

	/usr/src/build-extra/please.sh prerelease \
		--use-as-master-branch=mingit-2.14.x-releases \
		--only-mingit --no-upload --worktree=. HEAD

to use that MINGW-packages branch while building the mingw-w64-git
packages.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>