pacman-helper: use gpg.program, if set
In an automated build, we sometimes set the Git config variable
gpg.program to a shell script that prevents interactive programs from
stopping the flow of the automation.

Let's make that possible in pacman-helper, too.

This fix was made necessary by the upgrade from GPG 1.4 to GPG 2.2 that
was forced upon us by the MSYS2 maintainers.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>