please.sh upgrade git-lfs: work around inconsistent release notes
Git LFS' release notes are apparently written manually, which is the
only explanation this maintainer can understand for the many
inconsistencies between releases, such as file names with "v" vs without
"v", binaries in the top-level directory of the .zip files vs in a
subdirectory corresponding to the particular Git LFS version, etc

One particularly annoying inconsistency is that the SHA-256 sums were
listed *above* the file names in v2.5.0, when they were listed below for
all previous releases. And for v2.5.1, either.

Let's detect that inconsistency, like all the other ones, and work
around it individually.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>