Skip gitweb in Git for Windows
Git for windows is intended to used by end users. GitWeb, however wants
to be used in conjunction with a web server such as Apache, and even
requires a Perl module to be installed in that web server.

It is almost guaranteed that nobody uses Git for Windows that way, so
this patch filters out GitWeb from the installer.

In reference to issue #1362 https://github.com/git-for-windows/git/issues/1362.

Signed-off-by: Johan Abildskov <randomsort@gmail.com>