installer/portable: exclude xsltproc.exe
It was reported that xsltproc does not work due to a missing .dll.

Rather than including that .dll (and possibly others), drop the xsltproc
executable when bundling installers, too, keeping in mind that we are
talking about *Git for Windows*, not a project providing all commands on
Windows that Linux users might be used to.

This closes https://github.com/git-for-windows/git/issues/1775

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>