please.sh upgrade git: create more complete build-extra.bundle
Nowadays, we create three commits in build-extra: one commit to mention
the new Git version in the release notes, one to finalize them to the
new Git for Windows version, and one to update the package-versions
files in versions/.

So storing only 2 commits in the build-extra.bundle is not enough.

For good measure, store 9 commits (to be on the safe side), also when
generating the bundle after finalizing the release notes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>