installer (custom editor): hint at command-line options
Many editors on Windows are called through `.exe` files that return
immediately, without waiting for the editor to close. This is not what
Git needs: it needs the process to exit only when the file in question
has been edited. To this end, editors often support options to enable
that behavior, e.g. Visual Studio Code's `--wait` option.

To support this, we allow specifying options after the path to the
editor, but there was no indication so far. Up until now! :-)

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>