installer (custom editor): allow the path to be quoted
The convention when calling a command with options where the path to the
command contains spaces is to quote the path, e.g.

  "C:\Users\me\AppData\Local\Programs\Microsoft VS Code\Code.exe" --wait

However, the current code expects it to be specified without quotes:

  C:\Users\me\AppData\Local\Programs\Microsoft VS Code\Code.exe --wait

As it is not exactly clear in which form we want the command-line to be
specified, let's allow both forms.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>