installer: fix indentation
The majority of the lines in the `install.iss` script are indented using
spaces, using 4 spaces per indentation levels.

Over the course of the years, this has not exactly been maintained and a
few changes crept in that are inconsistent with this practice.

Let's make everything consistent again, and while at it, adjust the
.gitattributes accordingly.

Note: there was one stray, lonely semicolon that this patch also
deletes, which is not exactly an indentation fix, but it was indented
using tabs, too, so...

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>