installer (custom editor): allow testing the editor setting
It would be a bit unfortunate to configure a custom editor, then install
all of Git for Windows, only to find out the next time the user wants to
e.g. commit something that the editor does not, in fact, work, e.g. when
it requires options to exit after the file in question was closed.

Let's offer a convenient method to test things by adding a button to
edit a test file. Whether everything works as Git expects, or whether
anything went wrong, we'll tell the user about it (and enable/disable
the "Next" button accordingly).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>