shears --merging: need to rebase cousins
The problem with the merge commit starting the merging-rebase is that it
tells Git that the cousins do not need to be merged again, as they are
already part of the history. To work around that, let's just rebase
them, which generates new commits with new merge bases and everything
works as intended.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>