installer/msi: deprecate Git CMD
Nowadays, our default is to add the `cmd` directory to the `PATH`, so
*any* CMD is "a Git CMD".

It is time to discourage users to use Git CMD, though, as it is too easy
to exploit the fact that in CMD, the current directory is *always*
prefixed implicitly to the `PATH`. That is, if a repository contains a
script `git.cmd` in the top-level directory, running `git log` in the
top-level directory will execute that script, and that behavior cannot
be prevented.

So let's deprecate Git CMD.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>