please.sh publish: work around a known node.exe issue
It gets to the point of being ridiculous how often node.exe is broken
due to some random dependency being renamed because of a version jump.

In this instance, it is libcares-2.dll no longer being present because
it is now called libcares-3.dll. The most ridiculous part is that it
seems that copying the file to its old name makes node.exe work again.

So let's do that for now, and let's start thinking about better ways in
the future. Maybe there is an official, portable Node that we can use
instead of the too-often-broken mingw-w64-nodejs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>