please.sh: keep certain files even if a Pacman upgrade would delete them
In the previous commit, we listed the DLLs that need to be copied back
after upgrading from OpenSSL v1.0.2p to OpenSSL v1.1.1a, in order to
keep the .exe files in Git for Windows' SDK in a working state.

In this commit, we make use of that list to copy said files before
upgrading Pacman packages in please.sh (the `sync` and `install`
commands, and implicitly the `upgrade` command, are affected), and to
move them back into their original place afterwards.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>