installer/portable/mingit: include the OpenSSL v1.0.2p DLLs
As pointed out in the previous commits: after upgrading to OpenSSL
v1.1.1a, a ton of .exe files won't work because they link to the OpenSSL
v1.0.2p DLLs (which have a different name than the v1.1.1a DLLs).

Until we are certain that we do not have any .exe/.dll files in any of
the packages we use to generate Git for Windows installers, let's just
include both v1.0.2p and v1.1.1a DLLs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>