installer/portable/mingit: exclude OpenSSL v1.1.1a's documentation
These are not intended for development, that's what Git for Windows' SDK
is for.

This reduces the size of MinGit from ~36MB down to ~24MB (it used to be
at 21MB, the difference is due to the need to ship both OpenSSL
v1.0.2p's and v1.1.1a's .dll files side by side, for now).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>