sdk installer: fix installation into a path with spaces
The setup-git-sdk.bat script was not prepared to work with an
installation location whose path contains spaces.

Rather than fixing each and every usage of %cwd%, let's just change the
current working directory.

This also addresses the long-standing bug that we did not, in fact,
initialize the Git worktree in %cwd% at all, but the current directory
(which means that we ran with %cwd% as the current directory all along,
so that new `@CD "%cwd%"` is actually a no-op).

This fixes https://github.com/git-for-windows/git-sdk-64/issues/6

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>