sdk-installer: original .dll files can live in libexec/git-core, too
Because of DLL search path issues, we sometimes hard-link (or copy) .dll
files from bin/ to libexec/git-core/, to guarantee e.g. that no
ssleay32.dll that was installed by somebody into C:\Windows\system32
will be picked up.

The SDK installer script was ill-prepared for that, and this commit
fixes that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>