please.sh test_remote_branch: be careful about unknown commits
Travis has this funny thing where it sometimes asks the Windows job to
test a commit that is no longer reachable from `pu`.

To help with that, we introduced some skipping logic that pretends to
pass without problems in such a scenario.

However, we should not do this for `master`. Travis now also (at least
sometimes) asks to test for a merge commit that is not reachable from
*anywhere*, so `master` was skipped by mistake.

Let's fix this.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>