ever-green: offer to integrate new changes into the ever-green branch
The first step of ever-green branches is to update them with regard to
their original branches.

Example: the ever-green branch shears/pu of Git for Windows reflects its
`master` (original branch) as if it were rebased to git.git's `pu`
(upstream branch). Whenever Git for Windows' `master` advances, we will
want to include those changes in shears/pu. This here commit implements
that.

Note: for simplicity's sake, this patch rebases cousins. Example:

	master:			shears/pu

	- A - B ---- D		- A' - B'
	    \      /
	     --- C

Clearly, A' and B' are already-rebased versions of A and B, and we will
want to rebase only C and D. Because we rebase cousins with this here
current patch, the result will look like this:

	- A' - B' ---- D'
		\    /
		  C'

This is obviously incorrect, but it is a little tricky to find the
correct base commit(s) for "cousins". So for simplicity's sake, we keep
that logic out of the current commit. The next commit will work on that,
though, so that we will end up with this instead:

	- A' - B' --- D'
	   \        /
	    ---- C'

Since we do not *actually* want to rebase the original branch, we only
generate the todo list as if we wanted to, and then start a rebase of
the ever-green branch and inject that previously generated todo list
instead.

After integrating ("cherry-picking", but with merge commits) the new
changes of the original branch into the ever-green branch, we want to
rebase result onto the upstream branch if that one changed.

Note that we cannot simply run those two rebases separately: We do
expect merge conflicts to happen, *in particular* in the new changes of
the original branch. Meaning that the first rebase might abort with an
error, the user will fix the merge conflict and continue the rebase, but
after that rebase finished, the ever-green.sh script is no longer in
charge and therefore cannot kick off the second rebase.

So we cannot simply run two rebases in this script. To help with that,
we use a (supported) trick: we inject an `exec` command (calling the
ever-green.sh script) at the very end of the first rebase, to generate
the todo list of the second rebase and to append it to the current todo
list. The interactive rebase will detect that the todo list changed, and
process the appended commands. That way, both rebases will be combined
into a single run, and all is dandy.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>