ever-green: handle merging-rebased branches
Git for Windows' `master` branch uses a trick to *both* rebase the
branch thicket to upstream versions *and* keep a fast-forward history.
That trick is a "merging-rebase", i.e. before starting the rebase, a
fake merge pulls in the pre-rebase commit history via a merge commit
that does not change the tree at all (i.e. `-s ours`, to include the
commit history without taking any changes).

This strategy makes it relatively simple to look for the base commit of
the branch thicket (simply by looking for that fake merge, identified by
the fixed-form commit message) as well as for the previous tip of the
original branch (which is the second parent of that fake merge).

The only tricky thing here is that we have to re-generate a new fake
merge, usually copy-editing the commit message of the previous fake
merge.

Oh, and we have to accommodate all kinds of commit ranges that now not
only need to mark `onto` as uninteresting, but also that fake merge
commit; Otherwise, we would try to rebase not only the current versions
of the patches, but *all previous* versions of those patches, too, and
also the ones dropped during past rebases.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>