Work around libidn2 v2.1.0->v2.1.1a upgrade issues
So in 84da7711d (please.sh sync: work around libidn2 upgrade renaming
the DLL, 2019-01-11), we worked around a sudden rename of the .dll file
to libidn2-4.dll.

Naturally, even MSYS2 now understands that that file was not in need of
a rename (breaking the .exe files linking to that .dll file in the
process), so naturally they renamed it *again*, *back to the old name*.

I am sure so glad that we went through that exercise.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>