ever-green continue-rebase: avoid infinite loop when the user wants to stop
When `git rebase --continue` fails, it might very well be due to the
same problem that stopped the rebase before. Let's not try to continue
infinitely in such cases.

To detect that scenario, look for the `stopped-sha` file, and for good
measure also let's make use of the `msgnum` file that is already used to
display the progress in the Bash prompt.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>