ever-green: keep authorship when continuing after merge conflicts
Originally, we called `git commit` if there were resolved merge
conflicts, simply to avoid showing the editor.

But the way we did it lost authorship, because only `git rebase
--continue` would actually read back the `author-script`.

So let's install another fake editor, this time detecting whether HEAD
is still the same as when `git rebase --continue` was called.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>