ever-green: when no new changes are there, rebase to upstream right away
Ever-green branches need to be rebased in two phases: first picking new
changes (if any) from the original branch, and then rebasing onto a new
tip of the upstream branch (if any).

If only the upstream branch has a new tip, but no new changes were added
to the original branch, then there is nothing we need to do in the first
phase. Rather than adding an `exec ... nested-rebase ...` line to the
todo list in that case (which will immediately generate a new todo list
and prepend it), let's get to the second phase immediately: generate and
prepend the new todo list before even opening the todo list in the
editor.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>