sdk cd <project>: only pull master when it is safe to do so
Earlier, we blindly pulled the `master` branch of the `origin` remote.
Now we only do that if there are no local uncommitted changes, and if
the local branch is called `master`. Otherwise, we'll skip that pull.

The rationale for the new behavior is convenience: if the user did not
make changes, and did not create a topic branch, they most likely want
to be up to date. Otherwise, they want their work left alone.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>