git-extra: when unshallowing build-extra, avoid using http.sslbackend
This step happens in the `pkgver` function, which is called with
`/usr/bin/` *before* `/mingw64/bin` in the PATH, meaning that not our
beautiful mingw-w64-curl is in effect, but the MSYS2 curl that can only
handle the openssl backend.

So let's avoid configuring the backend in that scenario altogether.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>